Changes to Deck, Card and cleaning .gitignore (#2)
This adds the shuffle method to the Deck class. The default seed has been set to the current system time, but can be configured by being passed as an argument.

Custom `__str__` and `__eq__` methods have been added to the Card class for comparison.

The .idea folder has been cleaned to exclude files that can be auto-generated by PyCharm.
The .gitignore file has been updated to reflect the above changes.